reorganize the `ResearchSection` into several files
