+++ frontend/public/locales/en/translation.json
@@ -482,9 +482,9 @@
     "theseDataArePublishedUnderODCBY": "These data are published under the terms of the Open Data Commons Attribution License <1>(ODC-BY 1.0)</1>.",
     "downloadTheDatabase": "Download the database",
     "accessTheCodeOnGitHub": "Access the code on GitHub",
-    "scientificLiterature": "Scientific literature",
     "research": "Research",
     "weHopeToContribute": "\"We want to contribute to make the large-scale algorithms deployed robust and beneficial to all humanity.\"",
+    "scientificLiterature": "Scientific literature",
     "visualizeTheData": "Visualize the data",
     "youCanQuicklyExploreEtc": "You can quickly explore our public database with our <1>Tournesol Data Visualization</1> appplication made with Streamlit.",
     "personCorrelationCoefficientMatrix": "Pearson correlation coefficient matrix of criteria scores."
