+++ frontend/public/locales/fr/translation.json
@@ -488,9 +488,9 @@
     "theseDataArePublishedUnderODCBY": "Ces données sont publiées sous la licence Open Data Commons Attribution License <1>(ODC-BY 1.0)</1>.",
     "downloadTheDatabase": "Télécharger la base de données",
     "accessTheCodeOnGitHub": "Accédez au code sur GitHub",
-    "scientificLiterature": "Littérature scientifique",
     "research": "Recherche",
     "weHopeToContribute": "« Nous voulons contribuer à rendre les algorithmes déployés à grande échelle robustes, et bénéfiques pour toute l'humanité. »",
+    "scientificLiterature": "Littérature scientifique",
     "visualizeTheData": "Visualisez les données",
     "youCanQuicklyExploreEtc": "Vous pouvez rapidement explorer notre base de données publique en utilisant notre application <1>Tournesol Data Visiualization</1> réalisée avec Streamlit.",
     "personCorrelationCoefficientMatrix": "Matrice des coefficient de corrélation de Pearson entre les scores des critères."
