+++ frontend/src/pages/home/videos/HomeVideos.tsx
@@ -11,10 +11,10 @@ import TitleSection from 'src/pages/home/TitleSection';
 import PollListSection from 'src/pages/home/PollListSection';
 import ComparisonSection from 'src/pages/home/videos/sections/ComparisonSection';
 import ExtensionSection from 'src/pages/home/videos/sections/ExtensionSection';
+import RecommendationsSection from 'src/pages/home/videos/sections/RecommendationsSection';
+import ResearchSection from 'src/pages/home/videos/sections/research/ResearchSection';
 import { DEFAULT_POLL_STATS, getPollStats } from 'src/utils/api/stats';
 import { PollStats } from 'src/utils/types';
-import RecommendationsSection from './sections/RecommendationsSection';
-import ResearchSection from './sections/ResearchSection';
 
 const HomeVideosPage = () => {
   const { t } = useTranslation();
